movie-playlist-BE/requirements.txt
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.0
asttokens==2.0.5
async-generator==1.10
attrs==21.4.0
backcall==0.2.0
beautifulsoup4==4.10.0
bleach==4.1.0
branca==0.4.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.2
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
dj-rest-auth==2.2.4
Django==3.2.7
django-allauth==0.50.0
django-cors-headers==3.12.0
django-extensions==3.1.5
django-rest-auth==0.9.5
django-rest-framework==0.1.0
django-seed==0.3.1
djangorestframework==3.13.1
drf-yasg==1.20.0
entrypoints==0.4
executing==0.8.3
Faker==13.11.1
folium==0.12.1.post1
fonttools==4.29.1
googlemaps==4.6.0
h11==0.13.0
idna==3.3
inflection==0.5.1
ipykernel==6.9.1
ipython==8.2.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
itypes==1.2.0
jedi==0.18.1
jellyfish==0.9.0
Jinja2==3.1.2
joblib==1.1.0
jsonschema==4.4.0
jupyter==1.0.0
jupyter-client==7.1.2
jupyter-console==6.4.0
jupyter-core==4.9.2
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
kiwisolver==1.3.2
MarkupSafe==2.1.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mistune==0.8.4
mpmath==1.2.1
nbclient==0.5.11
nbconvert==6.4.2
nbformat==5.1.3
nest-asyncio==1.5.4
nltk==3.7
notebook==6.4.8
numpy==1.22.4
oauthlib==3.2.0
outcome==1.1.0
packaging==21.3
pandas==1.4.1
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.1.1
prometheus-client==0.13.1
prompt-toolkit==3.0.29
psycopg2==2.9.3
pure-eval==0.2.2
pycparser==2.21
Pygments==2.11.2
PyJWT==2.4.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
pyrsistent==0.18.1
PySocks==1.7.1
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.1
pywin32==303
pywinpty==2.0.2
pyzmq==22.3.0
qtconsole==5.2.2
QtPy==2.0.1
regex==2022.1.18
requests==2.27.1
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
scikit-learn==1.1.1
scipy==1.8.1
seaborn==0.11.2
selenium==4.1.2
Send2Trash==1.8.0
six==1.16.0
sklearn==0.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.1
sqlparse==0.4.2
stack-data==0.2.0
sympy==1.9
terminado==0.13.1
testpath==0.6.0
threadpoolctl==3.1.0
toposort==1.7
tornado==6.1
tqdm==4.63.0
traitlets==5.1.1
trio==0.20.0
trio-websocket==0.9.2
uritemplate==4.1.1
urllib3==1.26.9
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.2
wordcloud @ file:///C:/Users/april/Downloads/wordcloud-1.8.1-cp310-cp310-win_amd64.whl
wsproto==1.1.0
xlrd==2.0.1
